+++ src/main/java/com/example/demo/module/order/proxies/OrderSP01WebServiceProxy.java
@@ -17,7 +17,7 @@ public interface OrderSP01WebServiceProxy {
 
     @PostMapping("/{order_id}")
     List<SP01OrderBean> updateOrder(@RequestHeader("x-api-key") String token,
-            @PathVariable(name = "order_id", required = true) Long order_id,
+            @PathVariable(name = "order_id", required = true) Integer order_id,
             @RequestBody OrderDetailDTO updateOrder);
 
     @PostMapping
@@ -29,5 +29,5 @@ public interface OrderSP01WebServiceProxy {
             @RequestParam(name = "limit", required = false) Integer limit);
 
     @GetMapping("/{id}")
-    SP01OrderBean getOrderById(@RequestHeader("x-api-key") String token, @PathVariable Long id);
+    SP01OrderBean getOrderById(@RequestHeader("x-api-key") String token, @PathVariable Integer id);
 }
